+++ benchmarks/reorder_msgs.cpp
@@ -264,6 +264,7 @@ void reorder_sends(ParMat<int>& A,
 int main(int argc, char* argv[])
 {
     MPI_Init(&argc, &argv);
+    MPIL_Init(MPI_COMM_WORLD);
 
     int rank, num_procs;
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
@@ -352,6 +353,7 @@ int main(int argc, char* argv[])
 
     MPIL_Comm_free(&xcomm);
 
+    MPIL_Finalize();
     MPI_Finalize();
     return 0;
 }
