+++ library/include/neighborhood/neighbor_locality.h
@@ -5,24 +5,7 @@
 #include <vector>
 
 #include "communicator/MPIL_Comm.hpp"
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-typedef struct _CommData
-{
-    int num_msgs;
-    int size_msgs;
-    int datatype_size;
-    int* procs;
-    int* indptr;
-    int* counts;
-    int* indices;
-    char* buffer;
-} CommData;
-
-  
+#include "comm_data.h"
 
 void map_procs_to_nodes(const int orig_num_msgs,
                         const int* orig_procs,
@@ -52,15 +35,8 @@ void update_global_comm(CommData* global_send_data,
 void form_global_map(const CommData* map_data, std::map<long, int>& global_map);
 void remove_duplicates(CommData* comm_pkg);
 
+void map_indices(CommData* idx_data, std::map<long, int>& global_map);
+void map_indices(CommData* idx_data, const CommData* map_data);
 
-void destroy_comm_data(CommData* data);
-/** @brief Sets the CommData::num_msgs to provided value */
-void init_num_msgs(CommData* data, int num_msgs);
-/** @brief Sets CommData::size_msgs and allocates CommData::indices for indexing messages **/
-void init_size_msgs(CommData* data, int size_msgs);
-
-#ifdef __cplusplus
-}
-#endif
 
 #endif
