+++ library/source/communicator/CMakeLists.txt
@@ -1,6 +1,6 @@
 target_sources(locality_aware PRIVATE 
-	global_comms.cpp
-	topology_functions.cpp
 	get_tag.cpp
+	global_comms.cpp
     initializers.cpp
+	topology_functions.cpp
 )
